refactor: file organization, and fixed changing the type to string when sending a report
