+++ src/app/questionnaire/layout.jsx
@@ -6,9 +6,9 @@ import { FormProvider, useForm } from "react-hook-form";
 import { useTheme } from "@emotion/react";
 import { Box, Hidden, Stack, Container } from "@mui/material";
 import { useInputData } from "@/context/InputDataContext";
-import FormStepper from "@/components/FormStepper";
-import FormMobileStepper from "@/components/FormMobileStepper";
-import FormNavigationButton from "@/components/FormNavigationButton";
+import FormStepper from "@/components/step-manager/Stepper";
+import FormMobileStepper from "@/components/step-manager/MobileStepper";
+import FormNavigationButton from "@/components/step-manager/NavigationButtons";
 import Building from "@/model/Building";
 import monthlyDurationIntervals from "@/model/reference-data/monthlyDurationIntervals";
 import CustomAppBar from "@/components/AppBar";
