+++ src/components/step-manager/NavigationButtons.jsx
